backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easy-nlp-translate==0.4.0",
    "fastapi>=0.115.12",
    "faster-whisper>=1.1.1",
    "ffmpeg-python>=0.2.0",
    "python-multipart>=0.0.20",
    "srt>=3.5.3",
    "uvicorn>=0.34.2",
]
